mycode/requirements.txt
google-generativeai>=0.3.0
groq>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.0.0
supabase>=2.0.0 